Subscriptions utility class and rx.subscriptions package
https://github.com/Netflix/RxJava/issues/173
