+++ rxjava-core/src/main/java/rx/operators/OperationCombineLatest.java
@@ -32,6 +32,7 @@ import org.mockito.InOrder;
 import rx.Observable;
 import rx.Observer;
 import rx.Subscription;
+import rx.subscriptions.Subscriptions;
 import rx.util.functions.Func1;
 import rx.util.functions.Func2;
 import rx.util.functions.Func3;
@@ -798,7 +799,7 @@ public class OperationCombineLatest {
             public Subscription subscribe(Observer<String> Observer) {
                 // just store the variable where it can be accessed so we can manually trigger it
                 this.Observer = Observer;
-                return Observable.noOpSubscription();
+                return Subscriptions.empty();
             }
 
         }
