+++ rxjava-core/src/main/java/rx/operators/OperationToObservableIterable.java
@@ -26,6 +26,7 @@ import org.mockito.Mockito;
 import rx.Observable;
 import rx.Observer;
 import rx.Subscription;
+import rx.subscriptions.Subscriptions;
 import rx.util.functions.Func1;
 
 /**
@@ -53,7 +54,7 @@ public final class OperationToObservableIterable<T> {
             }
             observer.onCompleted();
 
-            return Observable.noOpSubscription();
+            return Subscriptions.empty();
         }
     }
