+++ rxjava-core/src/main/java/rx/operators/OperatorToIterator.java
@@ -12,6 +12,7 @@ import rx.Notification;
 import rx.Observable;
 import rx.Observer;
 import rx.Subscription;
+import rx.subscriptions.Subscriptions;
 import rx.util.Exceptions;
 import rx.util.functions.Func1;
 
@@ -116,7 +117,7 @@ public class OperatorToIterator {
             public Subscription call(Observer<String> observer) {
                 observer.onNext("one");
                 observer.onError(new TestException());
-                return Observable.noOpSubscription();
+                return Subscriptions.empty();
             }
         });
